refactor and address comment
