+++ hedera-node/hedera-app/src/main/java/com/hedera/node/app/records/impl/BlockRecordInfoUtils.java
@@ -2,12 +2,12 @@
 package com.hedera.node.app.records.impl;
 
 import com.hedera.hapi.node.state.blockrecords.BlockInfo;
+import com.hedera.node.app.blocks.impl.BlockImplUtils;
 import com.hedera.node.app.spi.records.BlockRecordInfo;
 import com.hedera.pbj.runtime.io.buffer.Bytes;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
 import java.time.Instant;
-import org.hiero.base.crypto.DigestType;
 
 /**
  * A utility class that provides methods for getting information from the {@link BlockInfo} object in order to
@@ -16,7 +16,11 @@ import org.hiero.base.crypto.DigestType;
  * otherwise suitable for a class hierarchy. So, utility methods FTW!
  */
 public final class BlockRecordInfoUtils {
-    public static final int HASH_SIZE = DigestType.SHA_384.digestLength();
+    /**
+     * The size in bytes of a single SHA-384 block hash. Re-exported from {@link BlockImplUtils#HASH_SIZE}, the
+     * canonical (block-format agnostic) definition, to avoid churning the many existing references to this constant.
+     */
+    public static final int HASH_SIZE = BlockImplUtils.HASH_SIZE;
 
     private BlockRecordInfoUtils() {
         throw new UnsupportedOperationException("Utility Class");
@@ -53,38 +57,7 @@ public final class BlockRecordInfoUtils {
      */
     @Nullable
     public static Bytes blockHashByBlockNumber(@NonNull final BlockInfo blockInfo, final long blockNo) {
-        return blockHashByBlockNumber(blockInfo.blockHashes(), blockInfo.lastBlockNumber(), blockNo);
-    }
-
-    /**
-     * Given a concatenated sequence of 48-byte block hashes, where the rightmost hash was
-     * for the given last block number, returns either the hash of the block at the given
-     * block number, or null if the block number is out of range.
-     *
-     * @param blockHashes the concatenated sequence of block hashes
-     * @param lastBlockNo the block number of the rightmost hash in the sequence
-     * @param blockNo the block number of the hash to return
-     * @return the hash of the block at the given block number if available, null otherwise
-     */
-    public static @Nullable Bytes blockHashByBlockNumber(
-            @NonNull final Bytes blockHashes, final long lastBlockNo, final long blockNo) {
-        final var blocksAvailable = blockHashes.length() / HASH_SIZE;
-
-        // Smart contracts (and other services) call this API. Should a smart contract call this, we don't really
-        // want to throw an exception. So we will just return null, which is also valid. Basically, if the block
-        // doesn't exist, you get null.
-        if (blockNo < 0) {
-            return null;
-        }
-        final var firstAvailableBlockNo = lastBlockNo - blocksAvailable + 1;
-        // If blocksAvailable == 0, then firstAvailable == blockNo; and all numbers are
-        // either less than or greater than or equal to blockNo, so we return unavailable
-        if (blockNo < firstAvailableBlockNo || blockNo > lastBlockNo) {
-            return null;
-        } else {
-            long offset = (blockNo - firstAvailableBlockNo) * HASH_SIZE;
-            return blockHashes.slice(offset, HASH_SIZE);
-        }
+        return BlockImplUtils.blockHashByBlockNumber(blockInfo.blockHashes(), blockInfo.lastBlockNumber(), blockNo);
     }
 
     // ========================================================================================================
